本番環境でメールを送れるようにする必要がある

+++ db/migrate/20240501002728_create_products.rb
@@ -4,8 +4,8 @@ class CreateProducts < ActiveRecord::Migration[7.0]
   def change
     create_table :products do |t|
       t.string :name, null: false
-      t.decimal :price, precision: 10, scale: 2, null: false
-      t.decimal :discouted_price, precision: 10, scale: 2
+      t.integer :price, precision: 10, scale: 2, null: false
+      t.integer :discouted_price, precision: 10, scale: 2
       t.text :description
       t.string :sku, null: false
